+++ frontend/src/app/services/index.ts
@@ -4,6 +4,7 @@ import { BackendLinkService } from './backendservice/backend-link.service';
 import { AuthInterceptor } from './http-interceptors/auth-interceptor.service';
 import { TokenService } from './authService/token-service.service';
 import { ConfigService } from './config/config.service';
+import { SocksManagerService } from './socksManager/socks-manager.service';
 
 export const services = [
   HttpErrorService,
@@ -12,4 +13,5 @@ export const services = [
   BackendLinkService,
   AuthInterceptor,
   TokenService,
+  SocksManagerService
 ];
